backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
transformers==4.36.0
torch==2.9.0
huggingface-hub==0.19.4
requests==2.31.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
rasterio==1.3.8